// frontend/api/models/KeyPopulationInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type KeyPopulationInput = {
    elderlyId: string;
    category: string;
    reason?: string;
    level?: string;
    ownerDoctorId?: string;
    followUpCycleDays?: number;
    status?: string;
};